# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documents",
        "version": "1.0.0"
    },
    "paths": {
        "/api/member/register": {
            "post": {
                "tags": [
                    "Member"
                ],
                "summary": "Register new member",
                "operationId": "23b02f85af10fcd61d677acb52a0c55c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "personal_id",
                                    "firstname",
                                    "lastname",
                                    "username",
                                    "email",
                                    "phone_number",
                                    "password",
                                    "password_confirmation",
                                    "language",
                                    "accept_consent"
                                ],
                                "properties": {
                                    "personal_id": {
                                        "type": "string",
                                        "example": "1198899232111"
                                    },
                                    "member_type": {
                                        "type": "string",
                                        "example": "internal,external"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Member"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "K"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "member_K"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "member_K@hotmail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0980000011"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "P@ssw0rd"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "P@ssw0rd"
                                    },
                                    "language": {
                                        "type": "string",
                                        "format": "language",
                                        "example": "THA"
                                    },
                                    "accept_consent": {
                                        "type": "string",
                                        "format": "accept_consent",
                                        "example": "T"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Member registered successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "username": {
                                                    "type": "string",
                                                    "example": "member_K"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "member_K@hotmail.com"
                                                },
                                                "create_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-15 14:23:45"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Register failed"
                    }
                }
            }
        },
...
    },
...
}